requirements.txt
beautifulsoup4
chromadb
faiss-cpu
langchain
langchain-openai
langchain-chroma
langchain-community
langchain_experimental
langchain_postgres
numexpr
numpy
openai
pandas
pgvector
psycopg
psycopg2
psycopg2-binary
pymupdf
pypdf
PyPDF2
python-dotenv
scikit-learn
streamlit
sqlalchemy
tiktoken
watchdog